COMBINED WEIGHT AND BODY FAT GRAPH:
Implemented a body fat graph to fit on top of the weight graph, using a
graph and series builder pattern, Combined them in to one graph and set
that next to a combined set of controls. Combined all of this with the
table to make the new weigh in view.