TESCO FOOD ITEM IMPORT PROTOTYPE HOOKED UP:
Very basic system hooked up allowing tesco items to be imported,
defaulting to 0.0 values if no information present, but using the
information if available. Imported in to database using the name - if it
exists it is not duplicated - references to follow later. Still very
flakey but can avoid using if needed in release and no image loading
enabled for performance concerns.