NUTRITIONAL UNIT:
Simple change to add Nutrition object and units in to model to allow
further units to be added easily and seamlessly in future. This caused
lots of deprecation. Alongside this a huge refactor of tables,
controllers, etc to use a builder pattern allowing a much easier
construction of tables and controls, also removing some rather pointless
but simple classes.