TABLE CONFIGURATION REFACTOR:
Introducing configuration for tree tables highlighted a huge gap in the
link between tree tables and table views in that there is no interface
but so much shared functionality. I have introduced separation
configurers to handle the common concepts specifically for each type but
this is possibly a risky change. All appears to work correctly but
approach with caution.